src/main/java/com/example/userstdd/usuarios/Usuarios.java
package com.example.userstdd.usuarios;

import java.util.ArrayList;

public class Usuarios {
    ArrayList<Usuario> usuarios = new ArrayList<>();
    public Usuario crear(Usuario usuario){
        usuarios.add(usuario);
        return usuario;
    }

    public String buscar(Usuario usuario){
        for(Usuario usuarioActual: usuarios){
            if(usuarioActual.nombre.equalsIgnoreCase(usuario.nombre)){
                return usuarioActual.nombre;
            }
        }
        return null;
    }

    public boolean borrar(Usuario usuario){
        return usuarios.remove(usuario);
    }
}
